% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assign.names.R
\name{assign.names}
\alias{assign.names}
\title{Use to attach column names (human-readable organism names) to a plain data.frame in accordance with the legend}
\usage{
assign.names(df, legend, sep = "X")
}
\arguments{
\item{\code{df}}{plain data.frame (presumably MLE data frame)}

\item{\code{legend}}{a legend data.frame (presumably from load.legend())}

\item{\code{sep}}{symbol that comes in front of a genome id in a column, default "X"}
}
\value{
same data.frame, but with attached column names (human-readable organism names)
}
\description{
Use to attach column names (human-readable organism names) to a plain data.frame in accordance with the legend
}
\examples{
legend<-load.legend(legend.file = "legend.csv")
mle.df<-extract.gprimes(mles.raw)
mle.df<-assign.names(mle.df,legend)
}

