% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/attach.genomeid.header.R
\name{attach.genomeid.header}
\alias{attach.genomeid.header}
\title{Use to attach marker symbols to numeric column names.}
\usage{
attach.genomeid.header(df, sep = "X")
}
\arguments{
\item{\code{df}}{data.frame, presumably RAW (!)}

\item{\code{sep}}{marker symbol to be attached, default is "X". Keep \code{sep} consistent throuhg analysis.}
}
\value{
the same data.frame as \code{df}, but with column names marked with \code{sep}
}
\description{
Use to attach marker symbols to numeric column names.
}
\examples{
master.table.raw<-read.bhs(bh.folder = "gBLASTer/bh")
master.table.raw<-as.data.frame(master.table.raw)
master.table.raw.10hcut<-restrict.minimal.hits<-function(df=master.table.raw)
master.table.raw.10hcut.header<-attach.genomeid.header(df=master.table.raw.10hcut)
}

