% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connect.derby.R
\name{connect.derby}
\alias{connect.derby}
\title{Connects to the gBLASTer embedded Derby database via JDBC.}
\usage{
connect.derby(db, derby.jar = "derby.jar", usr = "APP", pwd = "gblaster")
}
\arguments{
\item{\code{db}}{path to the debry database folder}

\item{\code{derby.jar}}{path to derby.jar library, default "derby.jar"}

\item{\code{usr}}{username, default "gblaster"}

\item{\code{pwd}}{password, default "gblaster"}
}
\value{
\code{connection} object, which represents the a connection to the database
}
\description{
Connects to the gBLASTer embedded Derby database via JDBC.
}
\examples{
db<-"/home/db/deby/"
derby.lib<-"/lib/derby.jar"
connect.derby(db, derby.jar=derby.lib, create=TRUE, usr="gblaster", pwd="gblaster")
}

