% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dist.spearman.one.R
\name{dist.spearman.one}
\alias{dist.spearman.one}
\title{Use to calculate distanse based on 1+spearman_correlation(each row in the dataframe)}
\usage{
dist.spearman.one(df)
}
\arguments{
\item{\code{df}}{data, can be data.frame or a matrix}
}
\value{
a dist object with the calculated values
}
\description{
Use to calculate distanse based on 1+spearman_correlation(each row in the dataframe)
}
\examples{
library(plyr)
library(dplyr)
library(tidyr)
bh.data<-read.bhs(bh.folder = "gBLASTer/bh",sep="_") \%>\%
 as.data.frame() \%>\%
 restrict.minimal.hits(minhit = 3) \%>\%
 normalize.scores() \%>\%
 attach.genomeid.header() \%>\%
 sign.bh.table()
data.3.6<-select.genomes(df = bh.data,g.ids = c(3,6))
}

