% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dist.stdev.one.R
\name{dist.stdev.one}
\alias{dist.stdev.one}
\title{Normalizes the rows by the standard deviation to make the more or less correlated rows appear closer}
\usage{
dist.stdev.one(df)
}
\arguments{
\item{\code{df}}{data, can be data.frame or a matrix}
}
\value{
the data, where each row is normalied by its standard deviation
}
\description{
Normalizes the rows by the standard deviation to make the more or less correlated rows appear closer
}
\examples{
library(plyr)
library(dplyr)
library(tidyr)
bh.data<-read.bhs(bh.folder = "gBLASTer/bh",sep="_") \%>\%
 as.data.frame() \%>\%
 restrict.minimal.hits(minhit = 3) \%>\%
 normalize.scores() \%>\%
 attach.genomeid.header() \%>\%
 sign.bh.table()
data.3.6<-select.genomes(df = bh.data,g.ids = c(3,6))
data.3.6.stn<-dist.stdev.one(data.3.6)
}

