% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getMLE.R
\name{getMLE}
\alias{getMLE}
\title{Use to calculate the maximum likelihood estimate of a hitset (presumably - from one genome)}
\usage{
getMLE(df, steps = 50, epsilon = 1e-06)
}
\arguments{
\item{\code{df}}{plain dataframe(no genome and names column) of raw score coordinates}

\item{maximum}{\code{steps} until the convergence of the ML, default 50 (more than enough actually)}

\item{\code{epsilon}}{minimum improvement in the coordinate after which the ML is considered to have converged, default 1e-6}
}
\value{
a \code{list} of \code{track} (a two column table, contains every stage of convergence process) and \code{gprime} (the actual ML estimate)
}
\description{
Use to calculate the maximum likelihood estimate of a hitset (presumably - from one genome)
}
\examples{
df<- read.bhs(folder) \%>\% data.frame()
MLE<-getMLE(df[,3:ncol(df)])
}

