% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getMLEs.R
\name{getMLEs}
\alias{getMLEs}
\title{Use to get a list of MLE values for a combined hit data.frame}
\usage{
getMLEs(df, steps = 50, epsilon = 1e-06)
}
\arguments{
\item{\code{df}}{a RAW or normalized RAW (has ORF names and ID_QUERY_GENOME column) data.frame}

\item{maximum}{\code{steps} until the convergence of the ML, default 50 (more than enough actually)}

\item{\code{epsilon}}{minimum improvement in the coordinate after which the ML is considered to have converged, default 1e-6}
}
\value{
a list of MLEs one per each genome
}
\description{
Use to get a list of MLE values for a combined hit data.frame
}
\examples{
master.table.raw<-read.bhs(bh.folder = "gBLASTer/bh")
data<-master.table.raw \%>\% data.frame() \%>\% restrict.minimal.hits(minhit = 40) \%>\% normalize.scores() \%>\% group_by(ID_QUERY_GENOME) \%>\% split(f=.$ID_QUERY_GENOME)
mles<-getMLEs(data)
}

