% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getRsquared.R
\name{getRsquared}
\alias{getRsquared}
\title{Use to get a pseudo Rsquared value for a logistic regression model}
\usage{
getRsquared(model)
}
\arguments{
\item{\code{model}}{logistic regression model object}
}
\value{
R squared value, calculated as 1-logLik(model)/logLik(null.fit), where null.fit is a model of cluster ~ 1
}
\description{
Use to get a pseudo Rsquared value for a logistic regression model
}

