% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hits.R
\name{load.hits}
\alias{load.hits}
\title{Use to load hits table}
\usage{
load.hits(hits.file)
}
\arguments{
\item{\code{hits.file}}{a path to the hits file}
}
\value{
a \code{dataframe} with the data
}
\description{
Use to load hits table
}

