% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logreg.mismatch.genes.R
\name{logreg.mismatch.genes}
\alias{logreg.mismatch.genes}
\title{Use to fit and minimize a logistic regeration model given the \code{data} and the \code{part} that will be used to train, \code{1-part} to predict and evaluate the model}
\usage{
logreg.mismatch.genes(data, org1, org2, num.bootstraps = 0, num.cpu = 1,
  cut = 0.5, part = 0.75)
}
\arguments{
\item{\code{data}}{a "coocked" data frame where colnames are as follows: X1..Xn, where n is the ids of genomes, rownames are: 1X0..mXn, where m is the largest orf_id}

\item{\code{org1}}{character name of the genome 1 (eg. X1)}

\item{\code{org2}}{character name of the genome 2 (eg. X2)}

\item{\code{num.bootstraps}}{number of bootstraps to run, default is 0, which means "do not run"}

\item{\code{num.cpu}}{number of cores to use for bootstrap, default is 1, makes sence only if \code{num.bootstraps} >0}

\item{\code{cut}}{probaility cut that allows to send a point either ot cluster 1 or cluster 2, defualt is 50%/50%, 0.5}

\item{\code{part}}{of the data to train the model, default is 0.75 (75%), leaving 0.25 to predict and evaluate the model}
}
\value{
a \code{list} train.data: a data.frame the was used to train the model,
working.data: a data.frame that was used to evaluate the model, fit: fitted and reduced minimal model itself, mismatch: a table of points that do not match clusters by the
perdiction, Rsq: r.squared of the model; in case \code{num.bootstraps}>0, the bootstrapped model also gets returned
}
\description{
Use to fit and minimize a logistic regeration model given the \code{data} and the \code{part} that will be used to train, \code{1-part} to predict and evaluate the model
}
\examples{
bh.data<-read.bhs(bh.folder = "/home/alext/Documents/Research/gBLASTer/bh",sep="_") \%>\%
as.data.frame() \%>\%
restrict.minimal.hits(minhit = 3) \%>\%
normalize.scores() \%>\%
attach.genomeid.header() \%>\%
sign.bh.table()

logreg.mismatch.genes(data = bh.data, org1 = "3", org2 = "6",part = 0.75)
}

