% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize.MLE.R
\name{normalize.MLE}
\alias{normalize.MLE}
\title{Use to normalize the MLE values if those were calculated on non-normalized RAW data}
\usage{
normalize.MLE(mle.df)
}
\arguments{
\item{\code{mle.df}}{a plain data.frame (no names and genome id columns).}
}
\value{
a data.frame that has each row normalized to [0,1] by the maximum value in the row (presumably "self hit MLE")
}
\description{
Use to normalize the MLE values if those were calculated on non-normalized RAW data
}
\examples{
mle.df<-extract.gprimes(mles.raw)
mle.df<-assign.names(mle.df,legend)
mle.df.norm<-normalize.MLE(mle.df = mle.df)
}

