% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.bhs.R
\name{read.bh.file}
\alias{read.bh.file}
\title{Reads a processed best hit file for a pair of genomes given (see params).
Will install dplyr, tidyr and data.table if not avaliable yet}
\usage{
read.bh.file(qgen.id, tgen.id, bh.folder, ending = "80.0.short", sep = "_",
  margin = "VS")
}
\arguments{
\item{\code{qgen.id}}{query genome id}

\item{\code{tgen.id}}{target genome id}

\item{\code{bh.folder}}{from a given bh folder}

\item{\code{ending}}{with a given file ending, defualt "80.0.short" which means "score cutoff 80"}

\item{\code{sep}}{specific character that separates the genome ids and the margin, default "_"}

\item{\code{margin}}{a margin symbol that separates the geniome ids, default "VS"}
}
\value{
a plain data.table representation of the file with "QUERY_ORF_ID", "COMULATIVE_BITSCORE", "ID_QUERY_GENOME" and "ID_TARGET_GENOME"
}
\description{
Reads a processed best hit file for a pair of genomes given (see params).
Will install dplyr, tidyr and data.table if not avaliable yet
}
\examples{
system.time(all.tables<-lapply(genome.ids, function(i){return(read.short.file(i[1],i[3],bh.folder))}))
}

