% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.bhs.R
\name{read.bhs}
\alias{read.bhs}
\title{Reads best hit files from the given folder and extracts a hit matrix (non-normalized)}
\usage{
read.bhs(bh.folder, ext = ".short", sep = "_")
}
\arguments{
\item{\code{bh.folder}}{that contains the shortened bh files}

\item{\code{ext}}{shortened files extension, default ".short"}

\item{\code{sep}}{separator symbol(s) that split the file names in query and target genomes}
}
\value{
a data.frame representation of the hit matrix where rows are orfs and columns are target genomes
}
\description{
Reads best hit files from the given folder and extracts a hit matrix (non-normalized)
}
\examples{
bh.folder<-"gBLASTer/bh"
head(test.table<-read.bhs(bh.folder = bh.folder))
}

