% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.orf.base.fasta.R
\name{read.orf.base.fastas}
\alias{read.orf.base.fastas}
\title{Use to read all files from a given folder and efficiently concatenate them into one large data.frame}
\usage{
read.orf.base.fastas(folder)
}
\arguments{
\item{\code{folder}}{that contains EXCLUSIVELY protein fasta records}
}
\value{
a data.frame that contains all sequences in one
}
\description{
Use to read all files from a given folder and efficiently concatenate them into one large data.frame
}
\examples{
fastas<-read.orf.base.fastas(folder = "gBLASTer/orfs")
}

