% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample.data.R
\name{sample.data}
\alias{sample.data}
\title{Gets a part of the data chosen randomly in ammounts indicated in \code{part}}
\usage{
sample.data(data, part = 0.75)
}
\arguments{
\item{\code{data}}{nonempty dataframe}

\item{\code{part}}{0.75 (75%) by default, the ammount of data to get randomly}
}
\value{
dataframe of randomly chosen part of the given dataframe
}
\description{
Gets a part of the data chosen randomly in ammounts indicated in \code{part}
}
\examples{
data.sampled <- sample.data(data=df, part=0.6)
}

