% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/connect.derby.R
\name{save.legend}
\alias{save.legend}
\title{Reads and returns the genome legend from the database.}
\usage{
save.legend(conn, out.file)
}
\arguments{
\item{\code{conn}}{connection to the gBLASTer database}

\item{\code{out.file}}{file to save the legend}
}
\description{
Reads and returns the genome legend from the database.
}
\examples{
conn<-connect.derby(db, derby.jar=derby.lib, create=TRUE, usr="gblaster", pwd="gblaster")
save.legend(conn$conn, "legend.txt")
}

