% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sign.bh.table.R
\name{sign.bh.table}
\alias{sign.bh.table}
\title{Attaches the query genome id to the dataframe (if data.table provided, will convert to data.frame).}
\usage{
\method{sign}{bh.table}(bh.table, sep = "X")
}
\arguments{
\item{\code{bh.table}}{bh data table that needs to have first two colums converted to rownames}

\item{\code{sep}}{separator symbol, default "X"}
}
\value{
data.frame with proper names attached
}
\description{
Attaches the query genome id to the dataframe (if data.table provided, will convert to data.frame).
}
\examples{
bh.folder<-"gBLASTer/bh"
test.table.sign<-sign.bh.table(bh.table = read.bhs(bh.folder))
}

