% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/to.bit.score.R
\name{to.bit.score}
\alias{to.bit.score}
\title{Converts a given raw BLAST score to bit score. Designed for protein alignments as by default
lambda and kappa are given for such.}
\usage{
to.bit.score(raw.score, lambda = 0.267, kappa = 0.041)
}
\arguments{
\item{\code{lambda}}{parameter, default 0.267 (as estimated by BLAST of 06.17.2015)}

\item{\code{kappa}}{parameter, default 0.041 (as estimated by BLAST of 06.17.2015)}
}
\value{
\code{bit score} for the given raw score and parameters
}
\description{
Converts a given raw BLAST score to bit score. Designed for protein alignments as by default
lambda and kappa are given for such.
}
\examples{
> to.bit.score(1263)
[1] 491.1153
}

