% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/to.raw.score.R
\name{to.raw.score}
\alias{to.raw.score}
\title{Converts a given bit BLAST score to raw score. Designed for protein alignments as by default
lambda and kappa are given for such.}
\usage{
to.raw.score(bit.score, lambda = 0.267, kappa = 0.041)
}
\arguments{
\item{\code{lambda}}{parameter, default 0.267 (as estimated by BLAST of 06.17.2015)}

\item{\code{kappa}}{parameter, default 0.041 (as estimated by BLAST of 06.17.2015)}
}
\value{
\code{raw score} for the given raw score and parameters
}
\description{
Converts a given bit BLAST score to raw score. Designed for protein alignments as by default
lambda and kappa are given for such.
}
\examples{
> to.raw.score(bit.score = 491.115)
[1] 1262.999
}

